requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.1
django-import-export==3.3.6
et-xmlfile==1.1.0
Flask==3.0.2
fonttools==4.47.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib==3.8.2
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
